Suppress one more Clang-Tidy warning [skip ci]

+++ simulator/infra/types.h
@@ -3,7 +3,7 @@
  * The types are used in algorithms affected by size of a variable
  *
  * @author Alexander Titov <alexander.igorevich.titov@gmail.com>
- * Copyright 2012 uArchSim iLab project
+ * Copyright 2012-2018 MIPT-MIPS project
  */
 
 // protection from multi-include
@@ -51,15 +51,11 @@ static const uint16 MAX_VAL16 = UINT16_MAX;
 static const uint32 MAX_VAL32 = UINT32_MAX;
 static const uint64 MAX_VAL64 = UINT64_MAX;
 
-
-
 #include <cassert>
 #include <iostream>
 
 class Cycle;
 class Latency;
-
-
 
 class Cycle
 {
@@ -95,15 +91,11 @@ class Cycle
         uint64 value;
 };
 
-inline auto operator""_Cl( unsigned long long int number)
+inline auto operator""_Cl( uint64 number)
 {
     assert( number <= MAX_VAL64);
     return Cycle( number);
 }
-
-
-
-
 
 class Latency
 {
@@ -139,11 +131,8 @@ inline auto operator""_Lt( uint64 number)
 
 inline auto operator*( uint64 number, const Latency& latency) { return latency * number; }
 
-
-
 Cycle   Cycle::operator+( const Latency& latency) const { return Cycle( value + latency.value); }
 Cycle   Cycle::operator-( const Latency& latency) const { return Cycle( value - latency.value); }
 Latency Cycle::operator-( const Cycle& cycle) const { return Latency( value - cycle.value); }
-
 
 #endif // #ifndef COMMON__TYPES_H
